requirements.txt
Django==4.2.4
djangorestframework==3.14.0
drf-yasg==1.21.7
mysqlclient==2.2.0
gunicorn==20.1.0
python-dotenv==1.0.0
django-cors-headers==4.2.0
sqlalchemy==2.0.20
PyJWT==2.8.0
loguru==0.7.2
urllib3==1.26.6
pandas==2.0.3
transformers==4.35.2
pytesseract==0.3.10
torch==2.0.1
bcrypt==4.1.2
ultralytics==8.0.238
pdf2image==1.16.3
paddleocr==2.7.0.3
paddlenlp==2.7.1
pinecone-client==3.0.1

## If Docker is not being utilized, kindly proceed to install the following libraries:
# paddlepaddle==2.5.2
# paddlepaddle-gpu==2.5.2
